Company_reputation_tracker_v2.2 - VIRKER!!/requirements.txt

dash>=2.9.0
dash-bootstrap-components>=1.4.0
plotly>=5.13.0
sqlalchemy>=2.0.0
textblob>=0.17.1
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
nltk>=3.8.1